sequencer: wire up "rewritten-hook" for REPLAY_HISTORY_EDIT
While the sequencer already knows to drive REPLAY_HISTORY_EDIT, we
currently skip the execution of the "rewritten-hook" as it is only
specific to interactive rebases. We do want to execute this hook though
for commits we're rewriting in the upcoming git-history(1) command.

Wire up the infrastructure so that we also execute this hook with
REPLAY_HISTORY_EDIT.

Signed-off-by: Patrick Steinhardt <ps@pks.im>
Signed-off-by: Junio C Hamano <gitster@pobox.com>