sequencer: allow callers to provide mappings for the old commit
When executing the "rewritten-hook" we provide a list of commit mappings
that tell the hook the original commit ID as well as the commit ID that
specific commit was rewritten to. Typically, these should always be
different from one another, as otherwise there wouldn't have been a
rewrite of those commits in the first place.

With two upcoming subcommands for git-history(1) that is not the case
though, as we have already written the new commits ahead of time. We
only use the sequencer infrastructure in that case to insert those
commits at the correct position in the graph. This has the consequence
that original and rewritten object IDs will be the exact same, which is
quite unhelpful.

Introduce infrastructure so that the caller can tell us the original
object ID for such already-rewritten objects.

Signed-off-by: Patrick Steinhardt <ps@pks.im>
Signed-off-by: Junio C Hamano <gitster@pobox.com>