reftable: check for trailing newline in 'tables.list'
In the reftable format, the 'tables.list' file contains a newline
separated list of tables. While we parse this file, we do not check or
care about trailing newlines. Tighten the parser in `parse_names()` to
return an appropriate error if there is no trailing newline.

This requires modification to `parse_names()` to accept a third argument
which will hold the error value.

Signed-off-by: Karthik Nayak <karthik.188@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>