xdiff: treat xdfile_t.rchg like an enum
Define macros NO(0), YES(1), MAYBE(2) as the enum values for rchg to
make the code easier to follow. Perhaps 'rchg' should be renamed to
'changed'?

A few of the code changes might appear to change behavior, such as:
    -    while (xdf->rchg[g->start - 1])
    +    while (xdf->rchg[g->start - 1] == YES)
because it appears the value of MAYBE is being ignored.  However, MAYBE
is only ever assigned as a value to a temporary array (dis1 & dis2) and
then as a last step use that temporary array to decide if it wants to
change xdfile_t.rchg[i] to YES or leave it as NO.  As such, rchg will
never have a value of MAYBE and thus there is no behavioral change.

Best-viewed-with: --color-words
Signed-off-by: Ezekiel Newren <ezekielnewren@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>