lib-gpg: allow tests with the GPGSM prereq first
When the 'GPG' prereq is lazily tested, `mkdir "$GNUPGHOME"` could
fail if the "$GNUPGHOME" directory already exists. This can happen if
the 'GPGSM' prereq has been lazily tested before as it uses
`mkdir -p "$GNUPGHOME"`.

To allow the GPGSM prereq to appear before the GPG prereq in some
test scripts, let's use `mkdir -p "$GNUPGHOME"` when the 'GPG' prereq
is lazily tested too.

This will be useful in a following commit.

Signed-off-by: Christian Couder <chriscool@tuxfamily.org>
Signed-off-by: Junio C Hamano <gitster@pobox.com>